reformats to App jsx
