+++ src/App.jsx
@@ -4,17 +4,35 @@ import Header from "./components/Header";
 import Departments from "./components/Departments";
 import Events from "./components/Events";
 import About from "./components/About";
+import { ThemeProvider, createTheme } from "@mui/material/styles";
+import { CssBaseline } from "@mui/material";
+import { amber } from "@mui/material/colors";
 
 const App = () => {
+  const getDesignTokens = (mode) => ({
+    palette: {
+      mode,
+      primary: {
+        ...amber,
+        ...(mode === "dark"),
+      },
+    },
+  });
+
+  const darkModeTheme = createTheme(getDesignTokens("dark"));
   return (
-    <BrowserRouter>
-      <Header />
-      <Routes>
-        <Route exact path="/" element={<Events />} />
-        <Route path="/departments" element={<Departments />} />
-        <Route path="/about" element={<About />} />
-      </Routes>
-    </BrowserRouter>
+    <ThemeProvider theme={darkModeTheme}>
+      <CssBaseline />
+
+      <BrowserRouter>
+        <Header />
+        <Routes>
+          <Route exact path="/" element={<Events />} />
+          <Route path="/departments" element={<Departments />} />
+          <Route path="/about" element={<About />} />
+        </Routes>
+      </BrowserRouter>
+    </ThemeProvider>
   );
 };
